EffectRefactoring step3

+++ Assets/_Content/Scripts/Effects/CartManipulation/ModifyCartAmount.cs
@@ -10,7 +10,7 @@ namespace Quackery.Effects
     {
         public override IEnumerator Execute(Effect effect)
         {
-            CartServices.AddToCartValue(effect.Value);
+            CartServices.AddToCartValue((int)effect.Value);
             yield return DefaultWaitTime;
 
         }
