+++ Assets/_Content/Scripts/Effects/PriceManipulation/PriceModifierBasedOnPriceEffect.cs
@@ -11,7 +11,7 @@ namespace Quackery.Effects
 
         public override IEnumerator Execute(Effect effect)
         {
-            DeckServices.BoostPriceOfCardsInHand(effect.Value, Card => Card.Price == _priceToMatch);
+            DeckServices.BoostPriceOfCardsInHand((int)effect.Value, Card => Card.Price == _priceToMatch);
             yield return DefaultWaitTime;
         }
     }
